requirements.txt
pip==21.3.1
Pillow==8.4.0
scrapy==1.5.1
requests==2.25.1
wget==3.2